requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
billiard==4.2.0
blis==0.7.11
catalogue==2.0.10
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.3
cron-descriptor==1.4.0
cymem==2.0.8
Django==4.2.7
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
gensim==4.3.2
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kombu==5.3.4
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
numpy==1.26.2
packaging==23.2
preshed==3.0.9
prompt-toolkit==3.0.41
# psycopg2==2.9.9
psycopg2-binary==2.9.9
PuLP==2.7.0
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.1
threadpoolctl==3.2.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
